+++ backend/src/baserow/contrib/database/table/usage_types.py
@@ -1,56 +1,68 @@
-from django.db.models import F, Func, Sum
-from django.db.models.fields.json import KeyTextTransform
+from django.db import connection
+from django.db.models import Sum
+from django.db.models.expressions import RawSQL
 
 from baserow.contrib.database.fields.models import FileField
+from baserow.contrib.database.models import Table
 from baserow.core.usage.registries import GroupStorageUsageItemType, UsageInBytes
 from baserow.core.user_files.models import UserFile
+
+FILENAMES_PER_GROUP_PLPGSQL_FUNCTION = """
+drop function if exists filenames_per_group(integer);
+create or replace function
+filenames_per_group(group_id integer) returns table(filename text)
+as
+$body$
+declare
+    field record;
+    filename text;
+begin
+for field in execute '{filefields_query}'
+loop
+    begin
+        return query execute 'select split_part(jsonb_array_elements(field_' || field.field_ptr_id || ') ->> ''name'', ''_'', 1) from {database_table_prefix}' || field.table_id;
+    exception
+        when undefined_table then
+            raise notice 'Could not find database_table_%', field.table_id;
+        when undefined_column then
+            raise notice 'Could not find field_% in {database_table_prefix}%', field.field_ptr_id, field.table_id;
+    end;
+end loop;
+end;
+$body$
+language plpgsql;
+"""
 
 
 class TableGroupStorageUsageItemType(GroupStorageUsageItemType):
     type = "table"
 
-    def calculate_storage_usage(self, group_id: int) -> UsageInBytes:
-        file_fields = FileField.objects.select_related("table").filter(
-            table__database__group_id=group_id,
-            table__database__trashed=False,
-        )
-
-        file_fields_by_table = {}
-        for file_field in file_fields:
-            table_id = file_field.table_id
-
-            if table_id not in file_fields_by_table:
-                file_fields_by_table[table_id] = [file_field]
-            else:
-                file_fields_by_table[table_id].append(file_field)
-
-        uniques = set()
-        for table_id in file_fields_by_table:
-            fields = file_fields_by_table[table_id]
-            table = fields[0].table
-            model = table.get_model(
-                fields=fields, field_ids=[], add_dependencies=False, use_cache=False
+    def register_plpgsql_functions(self):
+        # Using 9999 as placeholder for the group_id function argument
+        with connection.cursor() as cursor:
+            cursor.execute(
+                FILENAMES_PER_GROUP_PLPGSQL_FUNCTION.format(
+                    database_table_prefix=Table.USER_TABLE_DATABASE_NAME_PREFIX,
+                    filefields_query=str(
+                        FileField.objects.filter(
+                            table__trashed=False,
+                            table__database__trashed=False,
+                            table__database__group=9999,
+                        )
+                        .only("id", "table_id")
+                        .query
+                    ).replace("9999", "' || group_id || '"),
+                ),
             )
 
-            for field in fields:
-                unique_file_names_of_field = (
-                    model.objects.values(
-                        unique_file_names=KeyTextTransform(
-                            "name",
-                            Func(F(field.db_column), function="jsonb_array_elements"),
-                        )
-                    )
-                    .distinct()
-                    .values_list("unique_file_names", flat=True)
+    def calculate_storage_usage(self, group_id: int) -> UsageInBytes:
+        usage = (
+            UserFile.objects.filter(
+                unique__in=RawSQL(
+                    "select distinct filenames_per_group(%s)", (group_id,)
                 )
-                uniques_extracted = [
-                    unique_file_name.split("_")[0]
-                    for unique_file_name in unique_file_names_of_field
-                ]
-                uniques.update(uniques_extracted)
-
-        usage = UserFile.objects.filter(unique__in=uniques).aggregate(sum=Sum("size"))[
-            "sum"
-        ]
-
+            )
+            .only("size")
+            .aggregate(sum=Sum("size"))["sum"]
+        )
         return usage or 0